make add-book a seperate component
