+++ src/Components/Books.js
@@ -1,21 +1,11 @@
-// import PropTypes from 'prop-types';
-import { useState } from 'react';
 import { useSelector, useDispatch } from 'react-redux';
-import { addedBook, removeBook } from '../redux/books/books';
+import { removeBook } from '../redux/books/books';
+import AddBook from './AddBook';
 
 export default function Books() {
-  const [title, settitle] = useState('');
-  const [author, setauthor] = useState('');
-
   // consuming state
   const books = useSelector((state) => state.books);
   const dispatch = useDispatch();
-  const id = Math.random().toString();
-
-  const handleSubmit = (e) => {
-    e.preventDefault();
-    dispatch(addedBook({ title, author, id }));
-  };
 
   const handleDelete = (id) => {
     dispatch(removeBook(id));
@@ -23,6 +13,7 @@ export default function Books() {
 
   return (
     <>
+      <div className="container"><AddBook /></div>
       <div className="container">
         <div className="my-4 border border-3 p-0">
           <div>
@@ -50,18 +41,6 @@ export default function Books() {
             ))}
           </div>
         </div>
-        <h2 className="fs-2 text">Add New Book</h2>
-        <form className="row g-3">
-          <div className="col-7">
-            <input type="text" className="form-control" value={title} onChange={(e) => settitle(e.target.value.slice(0, 1).toUpperCase() + e.target.value.slice(1, e.target.value.length).toLocaleLowerCase())} id="inputPassword2" placeholder="Book Title" />
-          </div>
-          <div className="col-3">
-            <input type="text" className="form-control" value={author} onChange={(e) => setauthor(e.target.value.slice(0, 1).toUpperCase() + e.target.value.slice(1, e.target.value.length).toLocaleLowerCase())} id="inputPassword2" placeholder="Book Author" />
-          </div>
-          <div className="col-2">
-            <button type="submit" className="btn btn-primary mb-3" onClick={handleSubmit}>Add Book</button>
-          </div>
-        </form>
       </div>
     </>
   );
